Changed parameres types in CalculateMaxCableLengthOnReel method

+++ Cables/CableCalculations.cs
@@ -12,10 +12,10 @@ namespace Cables
         /// <param name="lengthBetweenFlanges">Расстояние между щёками барабана, мм</param>
         /// <param name="deltaToEdge">Расстояние от края намотки до края барабана, мм</param>
         /// <param name="cableDiameter">Диаметр наматываемой заготовки, мм</param>
-        /// <returns>Максимальная длина кабеля заданного диаметра, которая влезет на барабан с заданными параметрами, м (округлённая до ближайшего целого)</returns>
-        public static int CalculateMaxCableLengthOnReel(int reelDiameter, int barrelDiameter, int lengthBetweenFlanges, int deltaToEdge, double cableDiameter)
+        /// <returns>Максимальная длина кабеля заданного диаметра, которая влезет на барабан с заданными параметрами, м</returns>
+        public static double CalculateMaxCableLengthOnReel(double reelDiameter, double barrelDiameter, double lengthBetweenFlanges, double deltaToEdge, double cableDiameter)
         {
-            return (int)Math.Round(Math.PI * lengthBetweenFlanges * (Math.Pow(reelDiameter - 2 * deltaToEdge, 2) - barrelDiameter * barrelDiameter) / (4000 * cableDiameter * cableDiameter));
+            return Math.Round(Math.PI * lengthBetweenFlanges * (Math.Pow(reelDiameter - 2 * deltaToEdge, 2) - barrelDiameter * barrelDiameter) / (4000 * cableDiameter * cableDiameter));
         }
     }
 }
